Pushed logic for adding new mac addresses down into their respective components when registering new osp nodes.
