+++ fusor-ember-cli/app/components/new-node-registration-mac-address.js
@@ -6,6 +6,6 @@ export default Ember.Component.extend({
   }),
 
   cssId: Ember.computed('index', function() {
-    return `newNodeManualMacAddressAddInput${this.get('index')}`;
+    return `${this.get('prefix')}ManualMacInput${this.get('index')}`;
   })
 });
