+++ fusor-ember-cli/app/models/openstack-deployment.js
@@ -6,6 +6,7 @@ import {
   EqualityValidator,
   NumberValidator,
   HostAddressValidator,
+  IpAddressValidator,
   IpSubnetValidator,
   CidrValidator,
   AllValidator
@@ -22,6 +23,13 @@ const FlavorValidator = AllValidator.extend({
   validators: [
     PresenceValidator.create({}),
     EqualityValidator.create({doesNotEqual: 'baremetal'})
+  ]
+});
+
+const PresentIpValidator = AllValidator.extend({
+  validators: [
+    PresenceValidator.create({}),
+    IpAddressValidator.create({})
   ]
 });
 
@@ -78,6 +86,7 @@ export default DS.Model.extend(ValidatedModel, {
     overcloud_ext_net_interface: PresenceValidator.create({}),
     overcloud_private_net: PresentCidrValidator.create({}),
     overcloud_float_net: PresentCidrValidator.create({}),
+    overcloud_float_gateway: PresentIpValidator.create({}),
     overcloud_password: PresenceValidator.create({})
   }),
