+++ fusor-ember-cli/app/templates/components/new-node-registration.hbs
@@ -1,9 +1,9 @@
-{{#pf-modal-base idModal='nodeRegistrationModal'
+{{#pf-modal-base idModal="nodeRegistrationModal"
+                 cssClass="node-registration-modal"
                  openModal=openModal
                  title=newNodeTitle}}
 
   {{#pf-modal-body}}
-  <Br />
     {{#if isStep1}}
       {{partial "new-node-registration-step1-body"}}
     {{else}}
