+++ fusor-ember-cli/app/templates/components/pf-modal-base.hbs
@@ -1,4 +1,4 @@
-<div class="modal fade" id="{{idModal}}" tabindex="-1" role="dialog">
+<div class="modal fade {{cssClass}}" id="{{idModal}}" tabindex="-1" role="dialog">
   <div class="modal-dialog">
     <div class="modal-content">
