+++ fusor-ember-cli/app/templates/new-node-registration-auto.hbs
@@ -1,4 +1,4 @@
-<form class="form-horizontal new-node-form-manual">
+<form class="form-horizontal new-node-form-auto">
 
   {{select-req-f  cssId="autoDriverInput"
                   label="Driver"
