+++ fusor-ember-cli/app/templates/new-node-registration-manual.hbs
@@ -36,7 +36,7 @@
             isRequired=true}}
 
   {{#each nodeInfo.macAddresses as |macAddress key| }}
-    {{new-node-registration-mac-address macAddress=macAddress macAddressValidator=macAddressValidator index=key}}
+    {{new-node-registration-mac-address macAddress=macAddress macAddressValidator=macAddressValidator prefix="newNode" index=key}}
   {{/each}}
 </form>
 <div class="row">
